% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshtdaccesstoken.R
\name{refreshtdaccesstoken}
\alias{refreshtdaccesstoken}
\title{Get Access Token from TD Ameritrade Via Stored Refresh Token}
\usage{
refreshtdaccesstoken(callbackurl, consumerkey, storedtoken)
}
\arguments{
\item{callbackurl}{The callback URL defined in the TD Ameritrade Application.}

\item{consumerkey}{The "@AMER.OAUTHAP" appended TD Ameritrade API Key (called the "Consumer Key").}

\item{storedtoken}{The most recently issued Refresh Token... if this function breaks, this Refresh Token must be reissued.}
}
\value{
A string of characters that TD Ameritrade requires as an "Access Token".
}
\description{
This function uses a callback URL, an API key, and a monthly generated Refresh Token (for example saved to an RDS file) to generate an Access Token using a TD Ameritrade REST API endpoint.
}
\details{
The Access Token generated is good for 30 minutes. An HTTP request's authorization header uses this access token [preceded by "Bearer "] to retrieve data from TD Ameritrade's REST API servers.

[First draft: Friday, August 18th 2023]
[Last update: Tuesday, September 12th 2023]
}
\examples{
refreshtdaccesstoken( "https://mitzvah.capital", paste0( TD$APIKEY, "@AMER.OAUTHAP" ), readRDS( file = "~/Documents/tdameritrade-refresh-token.rds" ) )

}
\references{
TD Ameritrade's OAUTH Documentation/Reference (\url{https://developer.tdameritrade.com/content/simple-auth-local-apps}), Post Access Token Generation Link (\url{https://developer.tdameritrade.com/authentication/apis/post/token-0})
}
\seealso{
\code{\link{refreshtdaccesstoken}}
}
\author{
munair, \email{munair@gmail.com}
}
\keyword{access,}
\keyword{ameritrade,}
\keyword{td,}
\keyword{token}
